Get filtered assets data based on provided criteria

This function loads and filters the assets data based on criteria such as country, sector, technology, and company name.

Usage

get_filtered_assets_data(
  input_path,
  country_iso2 = NULL,
  sector = NULL,
  technology = NULL,
  company_name = NULL
)

Arguments

input_path

The path to the input data directory containing the necessary files for the TRISK model.

country_iso2

A character vector of ISO2 country codes to filter the assets. Defaults to NULL.

sector

A character vector of sectors to filter the assets. Defaults to NULL.

technology

A character vector of technologies to filter the assets. Defaults to NULL.

company_name

A character vector of company names to filter the assets. Defaults to NULL.

Value

A list containing the filtered assets data and other input data required for the TRISK model.